Move Data Structs to Core Modules (#131)
* move structs to core

* update struct references

* update augmentation docs

* update example and tutorial imports

* update tutorials and models docs

* update structs docs

* move data batch import to local scope to improve cli speed

* remove incorrect data item typing
