+++ autrainer/core/structs/__init__.py
@@ -1,0 +1,5 @@
+from .abstract_data_struct import AbstractDataBatch, AbstractDataItem
+from .data_struct import DataBatch, DataItem
+
+
+__all__ = ["AbstractDataBatch", "AbstractDataItem", "DataBatch", "DataItem"]
