+++ autrainer/datasets/utils/__init__.py
@@ -1,9 +1,3 @@
-from .data_struct import (
-    AbstractDataBatch,
-    AbstractDataItem,
-    DataBatch,
-    DataItem,
-)
 from .dataset_wrapper import DatasetWrapper
 from .file_handlers import (
     AbstractFileHandler,
@@ -23,13 +17,9 @@ from .zip_download_manager import ZipDownloadManager
 
 
 __all__ = [
-    "AbstractDataBatch",
-    "AbstractDataItem",
     "AbstractFileHandler",
     "AbstractTargetTransform",
     "AudioFileHandler",
-    "DataBatch",
-    "DataItem",
     "DatasetWrapper",
     "IdentityFileHandler",
     "ImageFileHandler",
