+++ autrainer/optimizers/sam.py
@@ -3,7 +3,7 @@ from typing import Callable, Tuple
 import torch
 
 import autrainer
-from autrainer.datasets.utils import AbstractDataBatch
+from autrainer.core.structs import AbstractDataBatch
 from autrainer.models import AbstractModel
 from autrainer.models.utils import create_model_inputs
