+++ docs/source/usage/tutorials.rst
@@ -284,43 +284,38 @@ of a :ref:`dataset <datasets>` implementation.
 Advanced Data Pipelines
 ~~~~~~~~~~~~~~~~~~~~~~~
 
-To create data and model pipelines that go beyond the standard :class:`~autrainer.datasets.utils.DataItem` convention of using
-only :attr:`DataItem.features` as input to the model, first create a new :class:`~autrainer.datasets.utils.DataItem` struct
-decorated with :class:`dataclasses.dataclass`:
+To create data and model pipelines that go beyond the standard :class:`~autrainer.core.structs.DataItem` convention of using
+only :attr:`~autrainer.core.structs.DataItem.features` as input to the model,
+first create a new :class:`~autrainer.core.structs.DataItem` dataclass that includes a new parameter (e.g., :attr:`meta`):
 
 .. literalinclude:: ../examples/tutorials/multi_branch_data.py
    :language: python
    :caption: multi_branch_data.py
-   :lines: 14-19
+   :lines: 11-16
 
-and override :class:`~autrainer.datasets.utils.AbstractDataBatch`:
-
-.. literalinclude:: ../examples/tutorials/multi_branch_data.py
-   :language: python
-   :caption: multi_branch_data.py
-   :lines: 22-41
-
-Following that, inherit from :class:`~autrainer.datasets.utils.DatasetWrapper`
-to create a :class:`torch.utils.data.Dataset`
-that iterates over your data 
-and returns your custom :class:`~autrainer.datasets.utils.AbstractDataBatch`
-(here we simply replicate :attr:`features`
-as our auxiliary features):
+Next, override :class:`~autrainer.core.structs.AbstractDataBatch` to include the new :attr:`meta` parameter:
 
 .. literalinclude:: ../examples/tutorials/multi_branch_data.py
    :language: python
    :caption: multi_branch_data.py
-   :lines: 44-52
+   :lines: 19-38
 
-Subsequently, inherit from :class:`~autrainer.datasets.AbstractDataset`
-to create a dataset
+Following that, inherit from :class:`~autrainer.datasets.utils.DatasetWrapper` to create a :class:`torch.utils.data.Dataset`
+that iterates over your data and returns your custom :class:`~autrainer.core.structs.AbstractDataBatch`
+(here we simply replicate :attr:`features` as our auxiliary features):
+
+.. literalinclude:: ../examples/tutorials/multi_branch_data.py
+   :language: python
+   :caption: multi_branch_data.py
+   :lines: 41-49
+
+Subsequently, inherit from :class:`~autrainer.datasets.AbstractDataset` to create a dataset
 that instantiates your :class:`~autrainer.datasets.utils.DatasetWrapper`:
 
-
 .. literalinclude:: ../examples/tutorials/multi_branch_data.py
    :language: python
    :caption: multi_branch_data.py
-   :lines: 55-73
+   :lines: 52-70
 
 Next, create a :attr:`ToyMultiBranch-C.yaml` configuration file for the dataset in the :attr:`conf/dataset/` directory:
 
@@ -329,8 +324,7 @@ Next, create a :attr:`ToyMultiBranch-C.yaml` configuration file for the dataset 
    :caption: conf/data/ToyMultiBranch-C.yaml
    :linenos:
 
-Finally, you must inherit from :class:`~autrainer.models.AbstractModel` 
-and create a model **with a matching signature**,
+Finally, inherit from :class:`~autrainer.models.AbstractModel` and create a model **with a matching signature**,
 in its forward pass to access the :attr:`meta` parameter:
 
 .. literalinclude:: ../examples/tutorials/multi_branch_model.py
@@ -338,12 +332,7 @@ in its forward pass to access the :attr:`meta` parameter:
    :caption: multi_branch_model.py
    :lines: 6-26
 
-
-Next, create a 
-:attr:`ToyMultiBranchModel.yaml` 
-configuration file 
-for the model in the :attr:`conf/model/` directory:
-
+Next, create a :attr:`ToyMultiBranchModel.yaml` configuration file for the model in the :attr:`conf/model/` directory:
 
 .. literalinclude:: ../examples/tutorials/MultiBranchModel.yaml
    :language: yaml
